requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
dj-database-url
Pillow
python-decouple
drf-spectacular
drf-spectacular-sidecar